feat: use dynamic import in page
